[clean] Discord専用のモジュール(post)をmainに移動しました。
