+++ src/common.rs
@@ -1,19 +1,3 @@
-pub mod post{
-    use serenity::{
-        model::channel::Message,
-        prelude::*,
-        utils::MessageBuilder,
-    };
-    pub async fn post_message(ctx: &Context, msg: &Message, message_str: String) {
-        let response = MessageBuilder::new()
-        .push(message_str)
-        .build();
-        if let Err(why) = msg.channel_id.say(&ctx.http, &response).await {
-            println!("Error sending message: {:?}", why);
-        }
-    }
-}
-
 pub mod channel{
     use std::env;
     use serenity::{
